Remove a reducer test for initial state
I know this works, but cannot test it without refactoring lots of code.
However, with the current library I am using (except), you cannot easily
intercept the message and mock a return value. There is a method for spies,
called 'spyOn', but it'll only work with objects, not on exported methods.

Source:
https://github.com/mjackson/expect/issues/139