Clean up some things
The ToneMultiPlayer wasn't being used anywhere. It *could* be useful
if I wanted to have one mega-object where each buffer is a Sample.
However, I want to be able to mix sound defs, between TR808 samples
& the ones generated by ToneJS.

Sticking with TR808 as the default sound definitions. So, change the
default 'tone' properties on each sequence.