Fix mute bug
Paying attention to the component life cycle methods, I noticed something
was off when it came to running re-rendering.

Even though the value in Redux state was correct I was not passing any
value for 'mute' to a ToneSequence object. The default value is false.

I had considered a toggleMute method on the ToneSequence object and also
refactoring it to be a class with other prototype methods, but opted to
stick with the Redux pattern of replacing it entirely.