Unit test sound module & ToneJS wrapper interface
Use mock-require to get tests to run

Need to mock the WebAudio API. Tried several other things, including:
'web-audio-api' & 'web-audio-test-api' modules. Those were very helpful
but was still getting errors deep into the ToneJS source code.

ToneJS uses the WebAudio API, which lives in the browser.
Would prefer to have my tests run with Node and would prefer not to
restructure my tests to run in the browser, if its only benefit was to
grab the API from it.